+++ AAI-Final-Assignment-WinForms/Entities/StaticEntity.cs
@@ -11,7 +11,7 @@ namespace AAI_Final_Assignment_WinForms.Entities
     public class StaticEntity : BaseGameEntity
     {
         public StaticEntity(Vector2D pos, GameWorld world, float scale, int textureWidth, int textureHeight,
-            double radius) : base(pos, world, scale, textureWidth, textureHeight, radius)
+            float radius) : base(pos, world, scale, textureWidth, textureHeight, radius)
         {
         }
     }
