+++ AAI-Final-Assignment-WinForms/Entities/Witch.cs
@@ -14,8 +14,8 @@ namespace AAI_Final_Assignment_WinForms.Entities
     {
         private Vector2D desiredVertex = new Vector2D();
 
-        public Witch(Vector2D pos, GameWorld world, double scale, int textureWidth, int textureHeight, double mass,
-            double maxSpeed, double maxForce, double radius
+        public Witch(Vector2D pos, GameWorld world, float scale, int textureWidth, int textureHeight, float mass,
+            float maxSpeed, float maxForce, float radius
         ) : base(pos,
             world, scale, textureWidth, textureHeight, mass, maxSpeed, maxForce, radius)
         {
@@ -37,7 +37,7 @@ namespace AAI_Final_Assignment_WinForms.Entities
                 new Rectangle((int)Pos.X - (int)Radius / 2, (int)Pos.Y - (int)Radius / 2, (int)Radius, (int)Radius));
         }
 
-        public override void Update(double timeElapsed)
+        public override void Update(float timeElapsed)
         {
             if (World.GameGraph.MovePath != null && World.GameGraph.MovePath.Count > 0)
             {
